Update Function_Documentation.txt
Added the majority of functions, except for those of digital signatures as the interface has to be updated to be able to test that it works. Need to check if the pubkeys and privkeys are better off stored as integers of base 10 as most of the RSA functions lack the statement int(<key>, 16) as we are no longer converting hexadecimal to base 10.